fix(gcs): Use BigInt for types that need >53 bits
When compiled to JS, Dart `int` only has 53 bits of precision. This is not enough to accommodate all GCS integer types.

+++ packages/google_cloud_storage/lib/src/client.dart
@@ -82,7 +82,7 @@ final class Storage {
   /// [Requester Pays]: https://docs.cloud.google.com/storage/docs/requester-pays
   Future<void> deleteBucket(
     String bucket, {
-    int? ifMetagenerationMatch,
+    BigInt? ifMetagenerationMatch,
     String? userProject,
     RetryRunner retry = defaultRetry,
   }) async => await retry.run(() async {
@@ -181,7 +181,7 @@ final class Storage {
   Future<BucketMetadata> patchBucket(
     String bucket,
     BucketMetadataPatchBuilder metadata, {
-    int? ifMetagenerationMatch,
+    BigInt? ifMetagenerationMatch,
     // TODO(https://github.com/googleapis/google-cloud-dart/issues/115):
     // support ifMetagenerationNotMatch.
     //
@@ -270,7 +270,7 @@ final class Storage {
     String name,
     List<int> content, {
     ObjectMetadata? metadata,
-    int? ifGenerationMatch,
+    BigInt? ifGenerationMatch,
     // TODO(https://github.com/googleapis/google-cloud-dart/issues/115):
     // support ifMetagenerationNotMatch.
     //
